feat: Add transaction service with CRUD operations and Kafka integration
- Implemented transaction service with create, read, and list functionalities.
- Added support for deposit and withdrawal transactions.
- Integrated account validation and balance retrieval from external account service.
- Introduced Kafka producer for balance updates using Avro serialization.
- Created health check endpoints for service monitoring.
- Configured environment variables for database, server, JWT, and Kafka settings.
- Added middleware for logging, CORS, and JWT authorization.
- Established database connection with retry logic and auto-migration.
- Defined transaction models and repository for data persistence.
- Implemented service interfaces for better abstraction and testing.

+++ Account-Service/internal/account/models/account.go
@@ -49,7 +49,7 @@ type Account struct {
     CustomerID      uuid.UUID          `json:"customer_id" gorm:"type:uuid;not null;index"`
     AccountType     AccountType        `json:"account_type" gorm:"type:varchar(20);not null"`
     AccountNumber   string            `json:"account_number" gorm:"type:varchar(20);unique;not null;index"`
-    Balance         int64              `json:"balance" gorm:"type:bigint;default:0"` // Store in cents
+    Balance         float64              `json:"balance" gorm:"type:bigint;default:0"` // Store in cents
     Currency        string             `json:"currency" gorm:"type:varchar(3);default:'USD'"`
     Status          AccountStatus      `json:"status" gorm:"type:varchar(20);default:'active'"`
     CreatedAt       time.Time          `json:"created_at" gorm:"autoCreateTime"`
